Add condition icon to weather page
Getting the icon path and drawing the icon is controlled by the define DRAWWEATHERICON. 
Weather icons are retrieved each time the weather page is shown (not cached). 
These are transparent PNGs provided by weatherapi.com on a CDN.
PNG drawing is based on this example https://github.com/Bodmer/PNG_TEST_ONLY/tree/master/png_test_url_spiffs_2x (with a fix to support_functions.h to correctly handle the x,y placement of the png).
Also note that weatherapi.com doesn't refesh the weather when you request it. They just send the last cached report. So the is_day value is frequently still no (zero) even when it is daytime and this affects the correct icon. I added basic handling of this in getWeather that uses fixed times for day/night rather than using sunrise/sunset or anything else.