Bug fixes
Weather page display was being called in a loop. This didn't matter before but adding the web call to get the PNG obviously became a problem. 
Also cleaned this looping page display up for the other pages just for hygiene. As part of this changed the starting value of pageJustChanged to true so the clock page would display on load. 
Moved the load & display of the PNG to the end of the displayWeather code so the whole page renders and then fills in the icon, which is much better for the UX.
Minor pixel placement adjustments on the weather page for degree symbol and "F".