More fixes around weather icons
The WeatherAPI.com conditions icons in 64x64 contained several that the PNGLE lib saw as badly formatted. The 128x128's were fine so I reworked things to use that size. 
The icon file errors were actually leading to crashes, which were probably related to some memory leak in the error condition in the PNGLE lib. 
Also a few misc fixes building on the previous work to improve page refresh handling.